backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
pydantic>=1.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Database
sqlalchemy[asyncio]>=2.0.0
alembic>=1.10.0
aiosqlite>=0.19.0
greenlet>=2.0.0

# WebSockets
websockets>=11.0.0

# AI & ML
openai>=0.27.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# Code Analysis
esprima>=4.0.1
astor>=0.8.1

# Utilities
python-dateutil>=2.8.2
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Documentation
mkdocs>=1.4.2
mkdocs-material>=9.1.0

# PDF Generation
weasyprint>=60.0

# Testing
pytest-cov>=4.0.0
pytest-mock>=3.10.0
